+++ src/components/UI/icons/IconCheck.vue
@@ -1,0 +1,16 @@
+<template>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    viewBox="0 0 24 24"
+    stroke-width="2"
+    stroke-linejoin="round"
+    stroke-linecap="round"
+    stroke="currentColor"
+    height="24"
+    fill="none"
+    class="svg"
+  >
+    <path d="M5 13l4 4L19 7" stroke="white" stroke-width="2"></path>
+  </svg>
+</template>
