+++ src/components/UI/icons/IconLeft.vue
@@ -1,0 +1,14 @@
+<template>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    class="w-auto h-full opacity-100 object-stretch"
+    viewBox="0 0 487 487"
+  >
+    <path
+      fill-opacity=".1"
+      fill-rule="nonzero"
+      fill="#FFF"
+      d="M0 .3c67 2.1 134.1 4.3 186.3 37 52.2 32.7 89.6 95.8 112.8 150.6 23.2 54.8 32.3 101.4 61.2 149.9 28.9 48.4 77.7 98.8 126.4 149.2H0V.3z"
+    ></path>
+  </svg>
+</template>
