--- src/components/UI/icons/IconMinus.vue
@@ -1,14 +1,0 @@
-<template>
-  <svg viewBox="0 -960 960 960" xmlns="http://www.w3.org/2000/svg">
-    <path d="M200-440v-80h560v80H200Z"></path>
-  </svg>
-</template>
-
-<style scoped>
-svg {
-  fill: #5f5f5f;
-  width: 15px;
-  height: 15px;
-  transition: 0.1s ease-in-out;
-}
-</style>
