--- src/components/UI/icons/IconPlus.vue
@@ -1,14 +1,0 @@
-<template>
-  <svg viewBox="0 -960 960 960" xmlns="http://www.w3.org/2000/svg">
-    <path d="M440-440H200v-80h240v-240h80v240h240v80H520v240h-80v-240Z"></path>
-  </svg>
-</template>
-
-<style scoped>
-svg {
-  fill: #5f5f5f;
-  width: 15px;
-  height: 15px;
-  transition: 0.1s ease-in-out;
-}
-</style>
