+++ src/components/UI/icons/IconRight.vue
@@ -1,0 +1,14 @@
+<template>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    class="object-cover w-full h-full"
+    viewBox="0 0 487 487"
+  >
+    <path
+      fill-opacity=".1"
+      fill-rule="nonzero"
+      fill="#FFF"
+      d="M487 486.7c-66.1-3.6-132.3-7.3-186.3-37s-95.9-85.3-126.2-137.2c-30.4-51.8-49.3-99.9-76.5-151.4C70.9 109.6 35.6 54.8.3 0H487v486.7z"
+    ></path>
+  </svg>
+</template>
