+++ src/pages/CardDescription.vue
@@ -3,8 +3,7 @@ import { useRoute } from 'vue-router';
 import { useSneakersStore } from '../store/state';
 import { computed, onMounted } from 'vue';
 import ButtonAddItem from '../components/UI/Buttons/ButtonAddItem.vue';
-import ButtonPlus from '../components/UI/Buttons/ButtonPlus.vue';
-import ButtonMinus from '../components/UI/Buttons/ButtonMinus.vue';
+import BaseButton from '../components/UI/Buttons/BaseButton.vue';
 
 const sneakersStore = useSneakersStore();
 const route = useRoute();
@@ -51,19 +50,17 @@ onMounted(() => {
             class="flex pt-1 gap-3"
             v-if="card.isAdded && card.currentQuantity >= 1"
           >
-            <button-minus
+            <BaseButton
               @click="sneakersStore.itemQuantityDecrement(card.id)"
-              class="border rounded px-1 hover:bg-gray-200"
+              type="minus"
             >
-              -
-            </button-minus>
-            <b>{{ card.currentQuantity }}</b>
-            <button-plus
-              class="border rounded px-2 hover:bg-gray-200"
+            </BaseButton>
+            <b class="mt-1">{{ card.currentQuantity }}</b>
+            <BaseButton
               @click="sneakersStore.itemQuantityIncrement(card.id)"
+              type="plus"
             >
-              +
-            </button-plus>
+            </BaseButton>
           </div>
         </div>
       </div>
